+++ src/inmanta/app.py
@@ -146,6 +146,7 @@ def start_scheduler(options: argparse.Namespace) -> None:
         await server_db_connect()
         await a.start()
 
+    LOGGER.info("Agent with Resource scheduler starting now")
     setup_signal_handlers(a.stop)
     IOLoop.current().add_callback(start)
     IOLoop.current().start()
