+++ src/inmanta/deploy/scheduler.py
@@ -419,7 +419,7 @@ class ResourceScheduler(TaskManager):
             return False
         return not (current_environment.halted or current_agent.paused)
 
-    async def refresh_agent_state_from_db(self, name: str) -> None:
+    async def refresh_agent_state_from_db(self, name: Optional[str]) -> None:
         """
         Refresh from the DB (authoritative entity) the actual state of the agent.
             - If the agent is not paused: It will make sure that the agent is running.
@@ -427,17 +427,28 @@ class ResourceScheduler(TaskManager):
 
         :param name: The name of the agent
         """
+        if name is not None:
+            if name in self._workers:
+                await self._workers[name].notify()
+                return
 
-        if name in self._workers:
-            await self._workers[name].notify()
-            return
+            requested_agent = await data.Agent.get(env=self.environment, endpoint=name)
+            if requested_agent:
+                self._create_agent(agent=requested_agent.name)
 
-        requested_agent = await data.Agent.get(env=self.environment, endpoint=name)
-        if requested_agent:
-            self._create_agent(agent=requested_agent.name)
+            if name not in self._workers and requested_agent is None:
+                raise LookupError(f"The agent `{name}` does not exist!")
+        else:
+            # The None represent the all agents action
+            agents_to_wake_up = await data.Agent.get_list(env=self.environment)
+            for agent in agents_to_wake_up:
+                if agent.name == const.AGENT_SCHEDULER_ID:
+                    continue
 
-        if name not in self._workers and requested_agent is None:
-            raise LookupError(f"The agent `{name}` does not exist!")
+                if agent.name in self._workers:
+                    await self._workers[agent.name].notify()
+                else:
+                    self._create_agent(agent=agent.name)
 
     async def is_agent_running(self, name: str) -> bool:
         """
